replace fetchDistributedTrace for getTraceDocuments (get_traces tool)
