+++ x-pack/solutions/observability/plugins/observability_agent_builder/server/routes/ai_insights/apm_error/fetch_apm_error_context.ts
@@ -16,7 +16,7 @@ import type {
 import { getApmIndices } from '../../../utils/get_apm_indices';
 import { parseDatemath } from '../../../utils/time';
 import { getServiceTopology } from '../../../tools/get_service_topology/get_service_topology';
-import { fetchDistributedTrace } from './fetch_distributed_trace';
+import { getTraceDocuments } from '../../../tools/get_traces/get_trace_documents';
 
 export interface FetchApmErrorContextParams {
   core: ObservabilityAgentBuilderCoreSetup;
@@ -102,13 +102,40 @@ export async function fetchApmErrorContext({
   if (traceId) {
     const traceContextPromise = (async () => {
       const apmIndices = await getApmIndices({ core, plugins, logger });
-      return fetchDistributedTrace({
+      return getTraceDocuments({
         esClient,
-        apmIndices,
-        traceId,
-        start: parsedStart,
-        end: parsedEnd,
-        logger,
+        traceIds: [traceId],
+        index: [apmIndices.transaction, apmIndices.span, apmIndices.error].flatMap((pattern) =>
+          pattern.split(',')
+        ),
+        size: 100,
+        startTime: parsedStart,
+        endTime: parsedEnd,
+        fields: [
+          '@timestamp',
+          'service.name',
+          'trace.id',
+          'event.outcome',
+          'parent.id',
+          'processor.event',
+          'transaction.id',
+          'transaction.name',
+          'transaction.type',
+          'transaction.duration.us',
+          'span.id',
+          'span.name',
+          'span.type',
+          'span.subtype',
+          'span.duration.us',
+          'span.destination.service.resource',
+          'error.id',
+          'error.grouping_key',
+          'error.culprit',
+          'error.log.message',
+          'error.exception.message',
+          'error.exception.handled',
+          'error.exception.type',
+        ],
       });
     })();
 
@@ -117,10 +144,10 @@ export async function fetchApmErrorContext({
       start,
       end,
       handler: async () => {
-        const { traceDocuments, isPartialTrace } = await traceContextPromise;
+        const [trace] = await traceContextPromise;
         return {
-          isPartialTrace,
-          documents: traceDocuments,
+          isPartialTrace: trace.isTruncated,
+          documents: trace.items,
         };
       },
     });
@@ -129,7 +156,10 @@ export async function fetchApmErrorContext({
       name: 'TraceServices',
       start,
       end,
-      handler: async () => (await traceContextPromise).services,
+      handler: async () => {
+        const [trace] = await traceContextPromise;
+        return trace.services;
+      },
     });
   }
